# docs/03-diseno/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/registro": {
      "post": {
        "operationId": "DonadorController_registrar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeDto"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "donador"
        ]
      }
    },
    "/api/auth/registro/confirmar": {
      "post": {
        "operationId": "DonadorController_confirmar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeDto"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "donador"
        ]
      }
    },
    "/api/auth/donador/sesion": {
      "post": {
        "operationId": "DonadorController_iniciar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SesionDonadorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SesionDto"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "donador"
        ]
      }
    },
...
  },
  "info": {
    "title": "Acopio · API",
    "description": "API REST de Acopio. Errores con forma { estado, codigo, mensaje, detalles? }. Autenticación con Bearer: el token de POST /api/auth/sesion.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SesionDto": {
        "type": "object",
        "properties": {
          "expiraEn": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "usuario": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "username": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "nombre": {
                "type": "string"
              },
              "rol": {
                "type": "string",
                "enum": [
                  "ADMIN",
                  "OPERADOR",
                  "AUDITOR",
                  "RECEPTOR",
                  "DONADOR"
                ]
              }
            },
            "required": [
              "id",
              "username",
              "nombre",
              "rol"
            ]
          }
        },
        "required": [
          "expiraEn",
          "usuario"
        ]
      },
      "ErrorDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "codigo": {
            "type": "string",
            "description": "Código estable para la interfaz, por ejemplo ULTIMO_ADMIN"
          },
          "mensaje": {
            "type": "string",
            "description": "Mensaje en español, listo para mostrar"
          },
          "detalles": {
            "description": "En errores de validación, un mensaje por campo. En algunos errores de dominio, datos para la interfaz (SALDO_INSUFICIENTE trae { saldo })",
            "anyOf": [
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "campo": {
                      "type": "string"
                    },
                    "mensaje": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "campo",
                    "mensaje"
                  ]
                }
              },
              {
                "type": "object",
                "additionalProperties": {}
              }
            ]
          }
        },
        "required": [
          "estado",
          "codigo",
          "mensaje"
        ]
      },
...
      "RegistroDto": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "contrasena": {
            "type": "string",
            "minLength": 12,
            "maxLength": 256
          },
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          }
        },
        "required": [
          "correo",
          "contrasena",
          "nombre"
        ]
      },
      "MensajeDto": {
        "type": "object",
        "properties": {
          "mensaje": {
            "type": "string"
          }
        },
        "required": [
          "mensaje"
        ]
      },
      "ConfirmarDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "token"
        ]
      },
      "SesionDonadorDto": {
        "type": "object",
        "properties": {
          "correo": {
            "type": "string",
            "minLength": 3,
            "maxLength": 254
          },
          "contrasena": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          }
        },
        "required": [
          "correo",
          "contrasena"
        ]
      },
...
    }
  }
}